feat(pg_repl): handle incoming messages in tasks
First, I have switched to the pgoutput plugin as I think it is more
efficient than json parsing. Next, I have created message handler to
handle incoming changes in async (fire and forget) Tasks.

This Tasks approach is not entirely correct since we cannot ensure the
correct order of applied changes. I am already thinking of changing it
to GenServer with queue (we should have a look to GenState as well).