feat(pg_repl): convert handler to GenServer
GenServer primitive is handling message queues for each call to it.
Since, we spawn only a single process of PgHandler GenServer, we know
that replication messages are being processed in order (with the
assumption that message queue of GenServer is keeping the order of
messages... I did not find it explicitely in documentation, I guess we
will learn the hard way).

I use `cast` instead of `call` simply because `repl` process does not
need to wait until the message is processed.