feat(pg_handler): handle UPDATE messages
TupleData lenght is defined by a number specifying the number of bytes

Right now I don't know how to handle tuple data properly. The reason is
that `O` tuple data have the original data for given row (before
update). Since it is the whole row, we need to know how many columns
given table has (it is not encoded in the message).

That means that parsing requires the knowledge of table schema. That can
be obtained from the RELATION message. I have noticed that at the first
startup of the application (so after initialization of the DB
connection) first messages are RELATION messages explainning each column
of the table in separate message. That means that the next part before
continuing with UPDATE must be handling of RELATION messages.

The parsing of the UPDATE for now is working correctly, it just does not
parse the whole binary. Because of that I am happy putting it for review
to see how the parsing of TupleData can be done for other messages.

Relates: #2

---v2
Added `!` to singal that method can bang. :D

Changed `tuple_type` to bitstring to help elixir show that we expect
character.