feat(core): create core msg protocol
---v2
Behaviour and protocol must be implemented by all messages.
I don't really like the any typing. However, I am not sure at this point
how to show that the result should be struct (in both cases very
specific, still unable to specify). :D

---v3
As we have discussed with Filip, it does not make sense to have core
behaviour, so I have deleted it.