feat(pg_subscriber): create PG subscriber supervisor
The idea is that each subscriber will initialize its own Supervisor
tree. That way we can make sure that the replication client is
configured with the same handler and that they are intialized in the
correct order. It will basically make the configuration a bit more
straighforward.

Note, we need to rething the supervisor child strategy. Currently, if
you try to configure multiple PG subscribers the second one will not
initialize because of naming collisions for Repl and Handler servers.

Relates: #20