feat: use Registry to implement publisher broadcast
Processes can register themselves to the Registry. In our case, we need
to register publishers. The registry stores PIDs of registered servers
and can store some other information. E.g. information about which
function to trigger when calling `Registry.dispatch/3`. That is very
convenient way of defining a generic entry to publisher (publishers
define which function is called during dispatch; it just needs to follow
specification defined in the PublisherContract).