refactor(pg_subscriber): create high level error structure
As part of this commit I have refactored handling of INSERT to match
handling of other messages. To fulfill the alpha mileston the only
missing part in subscriber is handle of TRUNCATE message.

The idea of `handle_message` is that it returns publish with message
that is send to publishers, not_publish to signal that everything is
fine but publishers don't need any information and error to signal that
there was some error while handling the message.

For now I am not sure how to react to all errors. So I am just catching
any errors that happen during handling without any action. The next step
will be to refactor from_data! function to return ok or error.