refactor: handle core conversion error
I have updated core DELETE message attributes (instead of columns to
where) to match proper handling of the message.

For DELETE I am choosing to put to where condition only columns that are
in the primary key. The reason is that if the table has primary key, we
get what primary keys to delete. If the table does not have primary key,
Postgres flags all columns as primary key (in the Relation message).